requirements.txt
blurhash
boto3==1.35.99
confusable-homoglyphs
erlastic
fastapi
Flask
flask-cors
Flask-Limiter
gunicorn
lxml
nltk
numpy
numpy==1.26.4
onnxruntime
openai
Pillow
pillow-heif
psycopg[binary]
pydantic[email]
PyYAML
redis
regex
uvicorn
websockets
xmltodict
